Narrow Survey's exponents to int, halving the struct
A double's true-ulp exponent lives in [-1074, 1023] and its top in
[-1073, 1024], so int carries six orders of magnitude more range than
the format can produce. long long bought nothing and cost half the
struct:

  before   24 bytes   two long long, two bool, six bytes of padding
  after    12 bytes   two int, two bool, two bytes of padding

The comment claimed 1 KB per 64-column survey against a 33.6 MB matrix,
one part in 32768. That figure was arrived at by counting four fields
and calling them 16 bytes; sizeof said 24, so the real ratio was one
part in 21845. It is now 768 bytes and one part in 43690, and a
static_assert pins the size so the quoted number stays a checked fact
rather than a remembered one.

The device already staged these as int, so host and device now agree
field for field and nothing narrows at the boundary. reserve_from_extents
and require_fit take int for the same reason.

One behavioural change rather than a width change: accumulate_column
computed `sc.min_exponent + log2_scale` to a long long, which was a
widening add while the field was long long and becomes an int add once
it is not. log2_scale is caller-supplied, so that sum could overflow on
the way to the very kExponentLimit check that exists to reject it. Both
operands are now widened before the add.

reserve_for keeps its long long locals: it reduces over DoubleParts, not
over a Survey, so narrowing it is a separate question.

+++ include/cbfp/cuda_accumulator.hpp
@@ -198,9 +198,9 @@ private:
     void validate_host_survey(const double *b, std::size_t col_stride);
     void sync_descriptors();
     void harvest_occupancy();
-    void reserve_from_extents(const long long *low, const long long *high,
+    void reserve_from_extents(const int *low, const int *high,
                               const char *any, std::size_t count);
-    void require_fit(std::size_t j, long long min_exponent, long long max_top);
+    void require_fit(std::size_t j, int min_exponent, int max_top);
     void grow_column(std::size_t j, std::size_t needed);
     void rescale_column(std::size_t j, int new_exponent);
     void ensure_slots(std::size_t words);
